# swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Aplikacija nudi podporo pri popisu vrst po Sloveniji",
    "title": "Biolog API.",
    "contact": {
      "name": "David Rubin",
      "email": "david.rubin95@gmail.com"
    },
    "version": "1.0.0"
  },
  "host": "localhost:4000",
  "basePath": "/api/v1",
  "paths": {
...
    "/species": {
      "get": {
        "description": "Pridobi vse lokalno shranjene vrste",
        "tags": [
          "species"
        ],
        "operationId": "getSpecies",
        "responses": {
          "200": {
            "description": "species",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/species"
              }
            }
          }
        }
      },
      "post": {
        "description": "Ustvari nov zapis o podatkah neke vrste",
        "tags": [
          "species"
        ],
        "operationId": "createSpecies",
        "parameters": [
          {
            "x-go-name": "Payload",
            "name": "species",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/species"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "species",
            "schema": {
              "$ref": "#/definitions/species"
            }
          }
        }
      }
    },
...
    "/users": {
      "get": {
        "description": "Pridobi vse uporabnike, ki imajo vsaj 1 javno opazanje",
        "tags": [
          "users"
        ],
        "operationId": "getUsers",
        "responses": {
          "200": {
            "description": "user",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/user"
              }
            }
          },
          "400": {
            "$ref": "#/responses/"
          }
        }
      },
      "post": {
        "description": "Ustvari novega uporabnika v podatkovni bazi",
        "tags": [
          "users"
        ],
        "operationId": "createUser",
        "parameters": [
          {
            "description": "Podatki o uporabniku'",
            "name": "User",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/user"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "user",
            "schema": {
              "$ref": "#/definitions/user"
            }
          },
          "400": {
            "$ref": "#/responses/"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "species": {
      "description": "Podatki so vzeti in spletne strani GBIF, tudi ID",
      "type": "object",
      "title": "Species (lokalno shranjena vrsta)",
      "required": [
        "id",
        "species",
        "kingdom",
        "family",
        "class",
        "phylum",
        "order",
        "genus",
        "scientificName",
        "canonicalName",
        "conservationStatus"
      ],
      "properties": {
        "canonicalName": {
          "description": "Kanonicno ime za vrsto (samo latinsko ime)",
          "type": "string",
          "maxLength": 128,
          "x-go-name": "CanonicalName",
          "example": "Passer domesticus"
        },
        "class": {
          "description": "Ime razreda za vrsto",
          "type": "string",
          "maxLength": 64,
          "x-go-name": "Class",
          "example": "Aves"
        },
        "conservationStatus": {
          "description": "Stanje ogrozenosti vrste",
          "type": "integer",
          "format": "int64",
          "maximum": 10,
          "minimum": 1,
          "x-go-name": "ConservationStatus",
          "example": 8
        },
        "family": {
          "description": "Ime druzine za vrsto",
          "type": "string",
          "maxLength": 64,
          "x-go-name": "Family",
          "example": "Passeridae"
        },
        "genus": {
          "description": "Ime roda za vrsto",
          "type": "string",
          "maxLength": 64,
          "x-go-name": "Genus",
          "example": "Passer"
        },
        "id": {
          "description": "Enolicni identifikator za vrsto,\nkljuc je enak kot pri GBIF API",
          "type": "integer",
          "format": "int64",
          "x-go-name": "ID",
          "example": 5231190
        },
        "kingdom": {
          "description": "Ime kraljestva za vrsto",
          "type": "string",
          "maxLength": 64,
          "x-go-name": "Kingdom",
          "example": "Animalia"
        },
        "order": {
          "description": "Ime reda za vrsto",
          "type": "string",
          "maxLength": 64,
          "x-go-name": "Order",
          "example": "Passeriformes"
        },
        "phylum": {
          "description": "Ime debla za vrsto",
          "type": "string",
          "maxLength": 64,
          "x-go-name": "Phylum",
          "example": "Chordata"
        },
        "scientificName": {
          "description": "Znanstveno ime za vrsto (latinsko ime, vcasih skupaj z avtorjem)",
          "type": "string",
          "maxLength": 128,
          "x-go-name": "ScientificName",
          "example": "Passer domesticus (Linnaeus, 1758)"
        },
        "species": {
          "description": "Ime vrste",
          "type": "string",
          "maxLength": 64,
          "x-go-name": "Species",
          "example": "Passer domesticus"
        }
      },
      "x-go-name": "Species",
      "x-go-package": "github.com/rubinda/biolog"
    },
    "user": {
      "description": "Podatki o uporabniku so vzeti od zunanjega avtentikatorja",
      "type": "object",
      "title": "User (uporabnik nase aplikacijo)",
      "required": [
        "id",
        "displayName",
        "externalID",
        "givenName",
        "familyName",
        "email"
      ],
      "properties": {
        "displayName": {
          "description": "Prikazno ime za uporabnika",
          "type": "string",
          "maxLength": 64,
          "x-go-name": "DisplayName",
          "example": "David Rubin"
        },
        "email": {
          "description": "Elektronski naslov uporabnika",
          "type": "string",
          "format": "email",
          "maxLength": 128,
          "x-go-name": "Email"
        },
        "externalID": {
          "description": "ID zunanjega avtentikatorja za tega uporabnika",
          "type": "string",
          "maxLength": 255,
          "x-go-name": "ExternalID",
          "example": "8457232358972358923566"
        },
        "familyName": {
          "description": "Priimek uporabnika",
          "type": "string",
          "maxLength": 32,
          "pattern": "[A-Za-z]+",
          "x-go-name": "FamilyName",
          "example": "Rubin"
        },
        "givenName": {
          "description": "Ime uporabnika",
          "type": "string",
          "maxLength": 32,
          "pattern": "[A-Za-z]+",
          "x-go-name": "GivenName",
          "example": "David"
        },
        "id": {
          "description": "8 mestni ID dolocenega uporabnika",
          "type": "integer",
          "format": "int64",
          "maximum": 99999999,
          "minimum": 10000000,
          "x-go-name": "ID",
          "example": 10000000
        },
        "picture": {
          "description": "URL do prikazne slike uporabnika",
          "type": "string",
          "maxLength": 255,
          "x-go-name": "Picture",
          "example": "https://www.biologapp.com/static/img/david-profile.png"
        },
        "publicObservations": {
          "description": "Pove ali uporabnik dovoli javen dostop do svojih opazanj",
          "type": "boolean",
          "x-go-name": "PublicObservations"
        }
      },
      "x-go-name": "User",
      "x-go-package": "github.com/rubinda/biolog"
    }
  },
...
}